backend/tests/test_thumbnail_integration.py
"""
Integration tests for thumbnail upload and serving workflow.

This tests the complete flow from upload to display.
"""

import pytest
import os
import tempfile
from io import BytesIO
from PIL import Image


def create_test_image(format='JPEG', size=(100, 100)):
    """
    Create a test image in memory.
    
    Args:
        format: Image format (JPEG, PNG, GIF, WEBP)
        size: Image dimensions as tuple (width, height)
    
    Returns:
        BytesIO object containing the image
    """
    img = Image.new('RGB', size, color='red')
    img_io = BytesIO()
    img.save(img_io, format=format)
    img_io.seek(0)
    return img_io


def test_thumbnail_upload_and_storage():
    """
    Test that thumbnail upload creates a file with unique name and returns URL.
    """
    # Create a test image
    test_image = create_test_image('JPEG', (200, 200))
    
    # Simulate the upload process
    # In real implementation, this would call the Flask endpoint
    # For now, we just verify the logic works
    
    import uuid
    from datetime import datetime, timezone
    
    # Generate unique filename (same logic as backend)
    file_extension = 'jpg'
    unique_id = str(uuid.uuid4())
    timestamp = datetime.now(timezone.utc).strftime('%Y%m%d_%H%M%S')
    unique_filename = f"{unique_id}_{timestamp}.{file_extension}"
    
    # Verify filename format
    assert unique_filename.endswith('.jpg')
    assert '_' in unique_filename
    assert len(unique_filename) > 20  # UUID + timestamp + extension
    
    # Verify URL format
    thumbnail_url = f'/api/courses/thumbnails/{unique_filename}'
    assert thumbnail_url.startswith('/api/courses/thumbnails/')
    assert thumbnail_url.endswith('.jpg')


def test_thumbnail_path_storage_in_course():
    """
    Test that course creation stores thumbnail path correctly.
    """
    # Simulate course data with thumbnail
    thumbnail_url = '/api/courses/thumbnails/test-uuid_20240101_120000.jpg'
    
    course_data = {
        'title': 'Test Course',
        'description': 'Test Description',
        'category': 'Programming',
        'thumbnail': thumbnail_url
    }
    
    # Verify thumbnail is stored
    assert 'thumbnail' in course_data
    assert course_data['thumbnail'] == thumbnail_url
    
    # Simulate retrieval (round-trip)
    retrieved_thumbnail = course_data.get('thumbnail')
    assert retrieved_thumbnail == thumbnail_url


def test_thumbnail_fallback_logic():
    """
    Test that missing thumbnails fall back to default.
    """
    default_thumbnail = 'https://images.pexels.com/photos/1181677/pexels-photo-1181677.jpeg?auto=compress&cs=tinysrgb&w=400'
    
    # Test case 1: No thumbnail field
    course1 = {'title': 'Course 1'}
    thumbnail1 = course1.get('thumbnail') or default_thumbnail
    assert thumbnail1 == default_thumbnail
    
    # Test case 2: Empty thumbnail field
    course2 = {'title': 'Course 2', 'thumbnail': ''}
    thumbnail2 = course2.get('thumbnail') or default_thumbnail
    assert thumbnail2 == default_thumbnail
    
    # Test case 3: Valid thumbnail
    course3 = {'title': 'Course 3', 'thumbnail': '/api/courses/thumbnails/test.jpg'}
    thumbnail3 = course3.get('thumbnail') or default_thumbnail
    assert thumbnail3 == '/api/courses/thumbnails/test.jpg'


def test_multiple_uploads_generate_unique_filenames():
    """
    Test that multiple uploads generate unique filenames even with same original name.
    """
    import uuid
    from datetime import datetime, timezone
    import time
    
    filenames = []
    original_name = 'course-thumbnail.jpg'
    
    for i in range(5):
        # Generate unique filename
        file_extension = 'jpg'
        unique_id = str(uuid.uuid4())
        timestamp = datetime.now(timezone.utc).strftime('%Y%m%d_%H%M%S')
        unique_filename = f"{unique_id}_{timestamp}.{file_extension}"
        filenames.append(unique_filename)
        
        # Small delay to ensure different timestamps
        time.sleep(0.01)
    
    # All filenames should be unique
    assert len(filenames) == len(set(filenames))
    
    # All should have .jpg extension
    assert all(f.endswith('.jpg') for f in filenames)


if __name__ == '__main__':
    pytest.main([__file__, '-v'])
